refactor: 테스트 코드도 패키지 옮기기 상품 도메인 별로 패키지 옮기기
